requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
pandas==2.2.3
PyMuPDF==1.24.10
Pillow==11.0.0
requests==2.32.3
google-genai==0.8.0